Backend/api/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pg8000
bcrypt
PyJWT==2.8.0
pytest==8.2.0
httpx==0.27.0
pytest-asyncio==0.23.6
